Dpkg::Shlibs::Objdump: API overhaul
Move most of Dpkg::Shlibs::Objdump::parse and
the complete ::parse_dynamic_symbol to ::Object
since these both actually only work on a specific
object. Also divide the parse function into to
parts _read and _parse where the first is responsible
for calling objdump and the second for parsing the output.
(Or any other filehandle it gets).

This allows for easier testing of the parsing part without
having actually to run objdump on a file.

Also add some tests to 200_Dpkg_Shlibs to verify that the
module still works after all these changes.

I tried to adapt dpkg-shlibdeps to the new API but haven't
verified that it still works.