Fix parsing of shlibs files with respect to packagetype
An shlibs line without packagetype should be used as fallback in case we
don't find any dependency for that packagetype (udeb). So end the parsing
in extract_shlibs only if the packagetype is given and if it matches with
the current one. If the packagetype is not given, remember the dependency
but continue parsing the file to check if we don't have a more specific
dependency.