Add "-x<package>" option to dpkg-shlibdeps to exlude packages
dpkg-shlibdeps now supports "-x<package>" options that can be used to exclude
packages from generated dependencies. This is particalularly useful to
avoid self-dependencies when a package contains a binary and a library (without
requiring an shlibs.local file to override the usual shlibs file). It might
also be used to avoid other unwanted dependencies (use with care though).
Closes: #41907, #109954