dpkg-shlibdeps: look into package directories before the root filesytem
When dpkg-shlibdeps tries to find a library, it will scan
the same directories as ld.so but at various "virtual root" (the real root
and also the root of the build trees). The root of the built
packages should have precedence otherwise we might analyze outdated
files which are installed in the root filesystem and which would
be replaced by the package being built.