dpkg-gencontrol: use the new Dpkg::Deps module to rewrite the dependencies
Normal dependency fields are rewritten by order of importance and simplified
by taking into account the dependencies already parsed. The dependencies are
also sorted to have a more predictable output (which is easier to compare in
a debdiff output).

The other relation (non-dependency) fields are simplified as unions
instead of intersections.