Swedish translation update.
Regenerated script translations for all languages and updated Swedish
translation. Also updated the Swedish manual page translation now that
I have the translation for the strings that were added to the scripts
but not in the previous version of the scripts POT file. I did not
translate that last round to avoid having inconsistencies in the
translated strings.