Fix input semantics of Debian changelog parser
* scripts/changelog/debian.pl: Restore the old
default behaviour to parse STDIN. Also fix the -l
option to specify a label for the parsed input, but
the actual file. Still allow --file to be used for
that. Use the new inhandle option of parse() to pass
STDIN directly. This should be faster and removes
the implicit dependency on IO::String.