Set PERL_DL_NONLAZY to 1 in perl scripts called by configuration scripts
* scripts/dpkg-divert.pl, scripts/update-alternatives.pl,
scripts/install-info.pl, scripts/dpkg-statoverride.pl: Set
PERL_DL_NONLAZY to 1 so that the usage of Locale::Gettext doesn't
cause run-time failures when this perl binary module is not
compiled for the corresponding perl version (and this happens in
configuration scripts, in particular preinst, during upgrades).
See http://bugs.debian.org/479711 for details.