Handle correctly packages in triggers-awaited state during recovery
* src/packages.c (process_queue): packages which are in triggers-awaited
state without any corresponding triggers-pending can be configured and
can thus satisfy dependencies of other packages.

The core of the problem was that if dpkg is interrupted, you can have
the following situation:

    Package: a
    Triggers-Awaited: b

    Package: b
    Status: ... installed

This is as expected but I obviously hadn't properly tested the
recovery, which is to do a no-op trigger processing step for b.  The
machinery in packages.c would be unimpressed by b's appearance in
progress_bytrigproc.