dpkg-source/3.0 (quilt): always provide valid name in QUILT_SERIES
* scripts/Dpkg/Source/Package/V3/quilt.pm: run_quilt() can be run
when no quilt series file exist yet. In that case we have to
provide the name of the series file that we want to create (as the
quilt command might create it). Up to now
File::Spec->rel2abs(undef) returned the name of the current
directory and we improperly used that as value of QUILT_SERIES.
This regression has been introduced in 1.14.21 while trying to
to force usage of a specific series file by setting QUILT_SERIEs.