dpkg-shlibdeps: support RUNPATH exactly like RPATH
* scripts/Dpkg/Shlibs/Objdump.pm (_parse): Support RUNPATH exactly
like RPATH but taking precedence over it.

In reality RUNPATH is like RPATH but it's used after having looked in
directories listed in LD_LIBRARY_PATH while RPATH is used before (but only
if RUNPATH is not set). See dlopen and ld.so manual page. We ignore that
difference in the case of dpkg-shlibdeps as LD_LIBRARY_PATH is usually not
used except for cases where RPATH is not used when it really should have
been.