dpkg-shlibdeps: optimize by caching symbols files and objdump objects
* scripts/Dpkg/Shlibs/Objdump.pm (parse): Split into add_object()
and parse().
* scripts/Dpkg/Shlibs/SymbolFile.pm (merge_object_from_symfile):
New function to reintegrate symbols from previous symfiles.
* scripts/dpkg-shlibdeps.pl: Optimizes it by caching parsed
symbols files and objdump objects. This way neither of the
libraries or symbols files are parsed more than once.

This patch significantly improves performance of dpkg-shlibdeps bringing
it near to performance levels of versions prior to 1.14.8 without loosing
any of new functionally at all. Memory requirements are reduced too.

This patch SHOULD NOT change the end result of dpkg-shlibdeps. If it
does, it is a bug.

Improved-by: Raphael Hertzog <hertzog@debian.org>