Do not lose conffiles while replacing them on the same install run
Only do the ensure_pathname_nonexisting calls when we are not going to
return due to an already existing directory or a file not to be
overwritten.

Those calls are making sure that .dpkg-tmp and .dpkg-new pathnames do
not exist, and will remove them if they do, which messes with the way
conffiles are handled by keeping those and acting on them later on.

Closes: #80416, #513857