update-alternatives: better handle missing files
* scripts/update-alternatives.pl (rename_mv): Integrate logic to
not fail if ENOENT is the reason why the rename failed.

The underlying problem in the previous implementation is not
immediately obvious but rename() can fail for multiple reasons
at the same time. In #99870 we have a case where rename() fails
with EXDEV and the system("mv",…) call fails due to ENOENT. In that
case, rename_mv() fails but the check $! == ENOENT also fails
when in fact it should not because ENOENT is really the reason
why rename_mv() failed.

Verifying the existence of the source file first is a sanity check
that avoids running mv when we know that it's going to fail and clutter
the screen with an undesired error message.