Update debian/*.install to work with DH_COMPAT=6
Some manual page translations are currently not included since they are
too out of date. However dh_install with DH_COMPAT=6 will fail if it
has nothing to install for a given wildcard expression. So rewrite the
wildcards to return the translations and the original manual page in
a single expression.