update-alternatives: complete rewrite
Full rewrite of update-alternatives. It was badly needed so that
we can actually understand its behaviour and implement new features
on top of it. This commit doesn't add any new feature but still improves
the behaviour of --config: it has a nicer output (with choices sorted
alphabetically, closes: #437060) and it accepts a path as well as the
index of the choice (to make it easier to write non regression tests for
this action). The test-suite is adjusted accordingly.