update-alternatives: skip slave link if associated file doesn't exist
All slave links are optional in the sense that they are only installed
if the currently associated file does exist. This allows most
update-alternatives call to succeed even if the admin removed
documentation files or other optional files. Closes: #143701
A message is still displayed by default in this case to warn the
user that something uncommon has been detected.

Adjusted the test suite to verify this behaviour.