dpkg-shlibdeps: don't fail on binaries outside of a package's dir
Dpkg 1.15.0 introduced a regression by trying to deduce the target
directory in order to adjust some warnings. This check failed when
dpkg-shlibdeps was run on files not (yet) installed in the package's
directory (debian/package/). Fix that and emit a warning so that
maintainers are encouraged to analyze the binaries once they have
been installed in their target directory.