dpkg-shlibdeps: make shlibs.local override symbols files too
As discussed in http://lists.debian.org/debian-devel/2009/03/msg00428.html
dropping the shlibs.local package override for symbols files was not
necessarily a good idea. So this commit makes dpkg-shlibdeps skip
the symbols files for a library if it detects that the relevant dependency
is available in the shlibs.local file.