dpkg-source: add --skip-debianization extract option
Extracting a source package with --skip-debianization can be interesting
when you want to import a source package in a VCS and want to import
the upstream sources without having precise knowledge of the internals
of all source package formats. This option is only supported for the
source formats "1.0", "2.0" and "3.0 (quilt)" that handle clearly
identified upstream sources.