dpkg-shlibdeps: fix code to find out minimal version among all symbols
dpkg-shlibdeps did not always correctly initialize symbol-based
dependencies for libraries having symbols associated with a version "0".

“$minver ||= $sym->{minver};” changed $minver to the first non-null version
that followed a null version. Because the null version is evidently
selected as the minimal version but that code considers it unset and
replaces it with whatever it has.

Add a non-regression test for this part of the code.