dpkg-source: don't die on SIGPIPE of uncompressors
Modify Dpkg::Source::CompressedFile to not die when uncompressors
processes (gunzip, bunzip, etc.) are killed by SIGPIPE. Recent tar
versions close the pipe before having read everything sent by the
uncompressor process when they encounter the end of the tar file.
This was of course problematic for Dpkg::Source::Archive.
This is a regression compared to etch's dpkg-source which dealt with
SIGPIPE properly.

The Dpkg::Source::Compressor::wait_end_process() function had to be
extended to be able to forward options to Dpkg::IPC::wait_child().