dpkg-source: don't complain on binary files that are ignored
To avoid mistakes with "3.0 (quilt)" source packages, dpkg-source fails if
it finds binary files that have not been whitelisted in the debian
directory. Unfortunately it also fails on binary files that are ignored
and that will not be included in the debian tarball. This commit fixes
that although not completely.

The exclude patterns passed to tar --exclude are used to match filenames
and also full path names inside the tarball. This commit only adds support
for simple filename match. It is enough for all realistic cases.