Add lintian override for embedded-zlib
dpkg-deb links statically with zlib so that it can still work
even if the shared library is broken or unavailable. It's a facility
to be able to recover from difficult situations without having
to manually unpack debian packages.

See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=139320 for
one discussion about this.