dpkg-source: do not set arch:any in dsc on arch-restricted packages
dpkg-source was pretty liberal in setting 'Architecture: any' in the
dsc: as soon as there are two binary packages, one being arch:all and
one being arch-restricted, you get arch:any in the dsc. This is
incorrect because the package will only build on the autobuilders
if there are architecture-dependent binary packages available.

Thus this patch only produces arch:any if at least one binary package
specifies it. Otherwise it outputs the set of all binary architectures
(including arch:all).

Closes: #526617

Signed-off-by: Philipp Kern <pkern@debian.org>