update-alternatives: don't remove real files conflicting with link
When an alternative is installed and a real file is in conflict, the real
file is kept. When the same alternative is fully removed, the real file
should be kept according to the same logic (but currently it was removed).
This commit modifies update-alternatives to follow this expected
behaviour.

Also add a non-regression test.