dpkg-source: fix automatic patch registration in format 3.0 (quilt)
When the automatic patch is updated due to a second call to
dpkg-source, it's not always properly registered with quilt.
It must be applied with quilt itself otherwise quilt pop
then quilt push might fail when supplementary files have been modified
outside of quilt's knowledge.

Furthermore quilt import was not a good choice as it failed trying to copy
the patch over itself. The new approach is to revert the patch and let
quilt reapply it with quilt fold. It will go over the list of files
modified by the new patch and add those which were not yet part of the
previous patch (hence adding the required unmodified files in
.pc/<patch-name>/).

Based-on-patch-by: Goswin von Brederlow <goswin-v-b@web.de>