dpkg-buildpackage: allow options -nc and -S together
This combination is generally not recommended as you want to ensure
that a source package contains only source and not any intermediary files
left-over by the previous build. Nevertheless it can be useful from time
to time and instead of bluntly refusing, we now simply warn the user
and let the build proceed.

-nc still implies -b if nothing else has been specified as it's routinely
used to test if a fix is working by restarting a previously interrupted
build where it failed.