Remove deprecated status for substvars in dpkg-source and dpkg-genchanges
Even if there's no use at all for substvars or they are (currently)
undesirable making the source not easily reproducible for Debian and
most derivatives using the same build processesor, it might be
interesting for uses like templated source package generation and
similar, or other interesting ideas we might not have thought about.

Also the same logic we applied for the removal of the substvar support
could be applied to most of the “Build options” as well, as those do not
make sense in the normal build for official packages (like -U, or -D).

Changed the code to stop producing warnings, and remove deprecated notes
from man page.