dpkg-scanpackages: be less annoying with warning about overrides
If the user doesn't specify an override file, then he doesn't care
about the fact that all packages are missing from the (non-existing)
override file. Disable the warning in this case.

In other cases, when the override file is explicitely given, it's a
normal feature that the packages missing from it are listed; exactly like
other weirdness are reported (incorrect maintainer info, spurious override
entries).

Based-on-patch-by: Piotr Engelking <inkerman42@gmail.com>