dpkg-deb: Do not warn when parsing unknown fields with “Private-” prefix
Fields prefixed with “Private-” are intended to be used privately by
individuals, organizations or companies, and are expected to never be
standardized. This guarantees them to never collide with any official
field in the future. Because there's still the danger that different
organizations might end up using the same field name, those should
never be visible in any public repository.

Closes: #353040

Based-on-patch-by: Nils Rennebarth <nils.rennebarth@funkwerk-ec.com>