Remove invalid master alternatives on upgrade
update-alternatives was very laxist in its past and allowed things that
are now rightfully forbidden. For example, you can't use alternatives as
slave when it's a master alternative already. However there is sometimes
cruft files in /var/lib/dpkg/alternatives/ due to historical changes (or
bugs) that incorrectly trigger this check.

We avoid this problem by removing the conflicting files on upgrade.