Remove all invalid master alternatives found
Initially I wanted to cleanup only problematic cases thinking that there
would be few. But apparently there must be an old version of
update-alternatives that created improper entries for slave alternatives
in /var/lib/dpkg/alternatives so I do a full scan once when we upgrade
from a version older than 1.15.3 and hope that we can forget about those
upgrade issues after that.