Dpkg::Deps: fix dependency parsing code
Fix dependency parsing code in Dpkg::Deps to not accept "foo\nbar"
even if foo is valid. A regex was improperly matching "\n" as
end of string due to usage of the "m" modifier.

Also improve the warning displayed when the dependency can't be parsed.

Add a non-regression test to ensure that this problem doesn't come back.

Reported-by: Andrew Sayers <andrew-dpkg@pileofstuff.org>