Obsolete --forget-old-unavail
On parse mark not-installed leftover packages for automatic removal from
the database on next dump. The states we consider for removal are
want_purge, want_deinstall and want_hold. The latter being the ancient
default for not-installed packages.

This makes the --forget-old-unavail option not useful anymore, so switch
it to be a no-op, and target it for future removal as obsolescent.

Closes: #33394, #429262