dselect: Add builddir to the quoted include search path
The curkeys.h file gets generated at build time, and if we use an out
of tree build directory, it cannot be found in the current search path.
But we don't want to add the builddir to the system include search path
as it will give file collisions when builddir is the same as srcdir.
Hence we only add it for quoted includes.