Dpkg::Changelog: fix heuristic to identify correct start entry
The heuristic to identify the version number that is most approaching
was not working properly for the case where the version number was smaller
than the oldest version (for example "0"). Add a non-regression test for
this case.

Also add supplementary warnings message so that people notice when
the parameters are discarded.