Add a dpkg cron.daily script to backup the status file
Create /etc/cron.daily/dpkg to handle the backup of /var/lib/dpkg/status
in /var/backups. This is taken out of the cron package and need no
conflicts/breaks as the code does nothing if the current status file is
already backupped.

Moving it out of the cron package makes it more visible for users
inspecting the content of the dpkg package.

Based-on-patch-by: Leo 'costela' Antunes <costela@debian.org>