libdpkg: Allow pkg:arch syntax in package relationship fields
Implement support for the packagename:archname syntax in package
relationships, required for multiarch.

Also, the only architecture value currently allowed is “any”, consistent
with round one of <https://wiki.ubuntu.com/MultiarchSpec>. This may be
relaxed before the wheezy release to allow for arch-specific cross
dependencies in the next release, but should probably not be relaxed
for package generation in order to avoid accidental archive uploads of
uninstallable packages.

[guillem@debian.org:
 - Rearrange struct deppossi for better memory alignment.
 - Do not print the arch qualifier if arch_none instead of NULL.
 - Check against arch_wildcard instead of the literal string.
 - Check archlength == 0 instead of arch[0] being NUL. ]

Signed-off-by: Guillem Jover <guillem@debian.org>