Tightening the Breaks for the install-info transition
The versions of info-browsers that dpkg breaks are a bit too
loose in some cases, namely when it "Breaks: foobar (<= x.y-z)", where
x.y-z is the Lenny version of info-browser foobar.  The problem is that
there might be an update of foobar in Lenny, say to x.y-zlenny1, which
dpkg also needs to break.