Dpkg::Control: new object interface
The new object interface combines the old Dpkg::Fields::Object
and the old parsecdata funtion.

The new object can be used like a regular hash and yet it behaves
intelligently taking into account all the specifics of control
information. Fields are case-insensitive, the order in which they have
been read is kept for output stability.

The object uses operator overloading and an internal tied hash so that
it just does what you mean.