Integrate dpkg-ftp into dselect
It requires perl to work but don't add it as a dependency
but only as a suggests and instead have a nice error
message to invite the user to install the perl package.

Other minor changes include removal of latin1 characters,
moving the README.mirrors file inside the method directory
and renaming the Dpkg::Ftp module to Dselect::Ftp.