libdpkg: Move copyfileperms to non-static file_copy_perms
This functionality is also needed by the conffile handling code to
ensure that the merge output is stored in a file with the same
permissions as the original conffile, preventing the accidental
opportunity for unintended information disclosure.

Therefore the function is moved into a new library module (file.{c,h}),
and given an appropriate prefix. Note that some of the translatable
error messages have been modified as they would otherwise be misleading.

Signed-off-by: Guillem Jover <guillem@debian.org>