libdpkg: Fix varbufprintf to not grow indefinitely
Use the C99 semantics for vsnprintf to get the final formatted string
size, and then call it again to do the actual work. Use varbuf_grow with
the know size, instead of always calling varbufextend, regardles of
needing to extend the buffer, which was making it grow indefinitely.

This also should speed it up a bit by not requiring to call vsnprintf
several times depending on the length of the final string. Only two
calls are required now, always.