dpkg-scanpackages: add new --medium option
This feature is supported in the dpkg-scanpackages implementation provided
by dpkg-multicd. Unfortunately it was using the "-m" option which got
granted to --multiversion in the official dpkg-scanpackages.

This option will only be useful for people wishing to create Packages.cd
files for use by the multicd dselect access method. Closes: #402527