Dpkg::Changelog::Entry, Dpkg::Control::Changelog: new modules
Dpkg::Changelog::Entry is a somewhat generic representation of a
changelog entry. It offers an interface exporting the basic information
that we need to extract for dpkg-parsechangelog.

Dpkg::Changelog::Entry::Debian is the implementation of this
interface for the typical Debian changelog entry.

Dpkg::Control::Changelog is the object representing the
information exported by dpkg-parsechangelog. It starts empty
as a simple Dpkg::Control of type CTRL_CHANGELOG.

Those new modules are not yet used by the Dpkg::Changelog modules.