Move parse_changelog() to a separate module Dpkg::Changelog::Parse
Despite its name, this function does not actuallay parse anything by
itself, it just calls external parsers and returns their result. This
interface is largely used by other tools and is not tied with any of the
other Dpkg::Changelog modules so it's best kept separate.

Many scripts have to be updated to cope with the renaming.