Dpkg::Changelog: completely update the API
The previous API was too different from the other modules inside dpkg-dev.

Dpkg::Changelog now behaves as an array of Dpkg::Changelog::Entry
and the documentation explains how it can be derived to create a new
changelog parser.

The non-regression tests as well as the official Debian
changelog parser script had to be updated to cope with the new API.