Dpkg::Control::Hash: handle arbitrary field values
Checking the validity of new field values at run-time was very expensive.
In order to avoid this check, the design of the object has been changed to
accept arbitrary values, they are simply escaped as necessary during
output. The only data lost in a save()/load() cycle are the
trailing whitespaces on every line of all values.

This API change implies fixes in all scripts/modules reading/creating
values for multiline fields.

The non-regression test also had to be adjusted.