dpkg: Do not bogusly ignore find exit code 1
Commit 68b56dc4edaaf90f5311f07b38871ddd61b84a48 introduced the
PROCNOERR to try to handle and be able to ignore find exit code 1,
which supposedly happened (not any longer) when using the -follow
option and a dangling symlink was found.

But it only prevented the ohshit() inside checksubprocerr(), and the
return value was still -1, so it was not ignoring the exit code 1,
and still ohshit()ing in the caller.

As find does currently exit with a 0 code even when finding broken
symlinks, let's remove this piece of broken logic.