Dpkg::Version: accept invalid versions by default
Accepting invalid version is the right default behaviour because
you often have to mix empty strings and versions and it's nice
to not have to special case too much.

In particular, we should be able to do "$vobj eq ''" without encountering
an error. Ensure this behaviour is kept by adding proper non-regression
tests.

Also enable fallback behaviour for all operators which have not been
explicitely implemented. The boolean operator returns true if the
version is valid, false otherwise.

Update the rest of the code that was doing "Dpkg::Version->new($v) || $v"
precisely to be able to store a version even if it's invalid.