Dpkg::Control::Hash: Strip leading spaces in the first line of a field's value
The fields in a compliant control-style file are allowed to have values
starting with any number of spaces. But that only applies to the first
line on multi-line field values, so do not do the same on subsequent
lines as those are handled differently.

Regression introduced in commit 7d8d193332478de82e20ed5b49053378801d160f.

Closes: #557547