Dpkg::IPC::fork_and_exec(): put modified environment variables in error messages
The descriptive command-line given to wait_child() now also includes the
environment variables that have been set since they may affect the
behaviour of the executed command and are likely needed to reproduce the
problem.

With this change, the quilt failures within dpkg-source are better
identified: QUILT_PATCHES can be checked to know if it was working
on the main tree or the temporary one and QUILT_PATCH_OPTS is
seen by users who can then independently reproduce the failure (and not
wonder why they can't reproduce it).