libdpkg: Properly increment seen fields in a stanza
The current code was not incrementing the contents of the pointed area,
but the pointer itself, which was making the check for already seen
(duplicate) fields a no-op, and thus non-fatal, contrary to what was
intended originally.

This should not be a problem in general as the tools in charge of
generating binary packages error out in case of duplicate fields, so
there should not be many broken packages on the wild.

This was spotted by a run with the clang static analyzer.