Dpkg::Shlibs::Cppfilt: new interface to the c++filt utility.
This new module can be used to do symbol demangling with the c++filt
utility. For performance reasons, the c++filt program is run
in the background and used as a daemon: for each line written
it immediately outputs the demangled line.

The module also takes care of the c++filt process management
(forking and termination).

Signed-off-by: Modestas Vainius <modax@debian.org>
Signed-off-by: Raphaël Hertzog <hertzog@debian.org>