libdpkg: Properly handle read errors on (de)compression
Instead of using the error handling code, failed reads are being
treated as end of file. This applies only when using zlib and libbz2.
In practice it probably has not caused problems because I/O errors
are rare, and often the program at the other end of the pipe can
notice the pipe unexpectedly closing.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Guillem Jover <guillem@debian.org>