libdpkg: Let backends decide default compression level
When compressing packages with gzip or bzip2, the tradeoff is
clear: a better compression ratio for a distributed package is
generally worth spending some extra time at build time. Since
better compressed packages are not much more inconvenient to
decompress at all, dpkg defaults to the maximum compression level
and developers rarely need to override that default.

On the other hand, LZ77-based decompressors use more memory at
decompression time for more tightly compressed packages, so the
maximum compression level of ‘9’ (which uses more than 32 MiB
of memory to decompress) can be too high.

With this patch, instead of defaulting to ‘9’, the compress_filter()
function sets the level to the backend default stored in the compressor
structure if the level is ‘-1’. All backends currently default to ‘9’
for now; no change in behavior is intended.

Based-on-patch-by: Jonathan Nieder <jrnieder@gmail.com>