Add -q option to dpkg-gensymbols, -c0 never fails AND shows a diff.
The rationale behind this change that there should be a way to prevent
dpkg-gensymbols from failing if there are lost symbols but still show a diff at
the same time (which -c0 would not without this patch).

Imagine a source package which builds multiple library binary packages that are
prone to arch-specific failures (quite typical for C++). Then dh_makeshlibs would
stop at the first dpkg-gensymbols failure giving only one diff. So it should be
possible to get diffs for all libraries in the single buildd run regardless of
check failures (using -c0). The fact that dpkg-gensymbols will never fail is
not that important if maintainer is pretty sure that symbols might be lost only
due to e.g. unmarked optional symbols or different mangling on other arches.
When updating to a new upstream, maintainer can set
DPKG_GENSYMBOLS_CHECK_LEVEL=1 in the environment to temporary restore checks.